src/plugins/nonebot_plugin_myb_exchange/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-myb-exchange"
version = "1.0.0"
description = "Nonebot2米游币商品自动兑换插件"
authors = ["惜月 <277073121@qq.com>"]
readme = "README.md"
license = "AGPL"
keywords = ["nonebot", "nonebot2", "mihoyo"]
homepage = "https://github.com/CMHopeSunshine/LittlePaimon/tree/Bot/src/plugins/nonebot_plugin_myb_exchange"
repository = "https://github.com/CMHopeSunshine/LittlePaimon/tree/Bot/src/plugins/nonebot_plugin_myb_exchange"
documentation = "https://github.com/CMHopeSunshine/LittlePaimon/tree/Bot/src/plugins/nonebot_plugin_myb_exchange#readme"
packages = [
    { include = "nonebot_plugin_myb_exchange/*.py", from = ".."}
]

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.4"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-apscheduler = "^0.1.2"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
